requirements.txt
supabase
pandas
python-dotenv
numpy

